Reduce test of operators to actual operators
Since the project extended behind its initial goal and know include
aliases which are not only for strictly ideographic identifiers, some
tests where undully performed on methods for which they were not
expected.

This commit fix this by reducing the indentifiers considered ideographic
to those composed without a single alphabetic sign.

This fix all errors raised in previous commits, but it doesn't cover
tests of the excluded aliases, a work left for later.