Provides lexicalized alternatives to Array operators
This also required to treat similar methods, as the will to use names
with the morph -mix- arised.

Moreover, others objects where also impacted, as the will to avoid
introduction of exact same term for different purpose is taken into
consideration.

There is however still quite a bit of work to cover all Array with fixed
goal of having a trigraph and a -mix- word in many cases.